Split the parser.py file into io.py and entrypoint.py
